Updated the search fragment.
First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

First commit of the Liber app.

Finished the search fragment

Paging data refresh

Arranged the folders

Added a new drawable

Added a constant for retrofit